Deleting Extra Files
There were a ton of extra files, so I went through and deleted them.
