--- moviesstore/home/admin.py
@@ -1,4 +1,0 @@
-from django.contrib import admin
-from .models import Movie
-admin.site.register(Movie)
-# Register your models here.
