--- moviesstore/home/templates/movies/index.html
@@ -1,30 +1,0 @@
-{% extends 'base.html' %}
-{% block content %}
-{% load static %}
-<div class="p-3">
-  <div class="container">
-    <div class="row mt-3">
-      <div class="col mx-auto mb-3">
-        <h2>List of Movies</h2>
-        <hr />
-      </div>
-    </div>
-    <div class="row">
-      {% for movie in template_data.movies %}
-      <div class="col-md-4 col-lg-3 mb-2">
-        <div class="p-2 card align-items-center pt-4">
-          <img src="{% static 'img/about.jpg' %}"
-            class="card-img-top rounded">
-          <div class="card-body text-center">
-              <a href="{% url 'movies.show' id=movie.id %}"
-              class="btn bg-dark text-white">
-              {{ movie.name }}
-            </a>
-          </div>
-        </div>
-      </div>
-      {% endfor %}
-    </div>
-  </div>
-</div>
-{% endblock content %}
