--- moviesstore/home/templates/movies/show.html
@@ -1,19 +1,0 @@
-{% for review in template_data.reviews %}
-          <li class="list-group-item pb-3 pt-3">
-            <h5 class="card-title">
-              Review by {{ review.user.username }}
-            </h5>
-            <h6 class="card-subtitle mb-2 text-muted">
-              {{ review.date }}
-            </h6>
-            <p class="card-text">{{ review.comment }}</p>
-            {% if user.is_authenticated and user ==
-              review.user %}
-            <a class="btn btn-primary"
-              href="{% url 'home.edit_review'
-              id=template_data.movie.id
-              review_id=review.id %}">Edit
-            </a>
-            {% endif %}
-          </li>
-          {% endfor %}
