+++ moviesstore/home/views.py
@@ -3,10 +3,10 @@ from django.shortcuts import render
 def index(request):
     template_data = {}
     template_data['title'] = 'Movies Store'
-    return render(request, 'home/index.html', {
+    return render(request, 'home/templates/home/index.html', {
         'template_data': template_data})
 def about(request):
     template_data = {}
     template_data['title'] = 'About'
-    return render(request, 'home/about.html' ,
+    return render(request, 'home/templates/home/about.html',
                   {'template_data': template_data})
