+++ moviesstore/movies/templates/movies/edit_review.html
@@ -1,0 +1,27 @@
+{% extends 'base.html' %}
+{% block content %}
+<div class="p-3">
+  <div class="container">
+    <div class="row mt-3">
+      <div class="col mx-auto mb-3">
+        <h2>Edit Review</h2>
+        <hr />
+        <form method="POST">
+          {% csrf_token %}
+          <p>
+            <label for="comment">Comment:</label>
+            <textarea name="comment" required
+            class="form-control" id="comment">{{
+            template_data.review.comment }}</textarea>
+          </p>
+          <div class="text-start">
+            <button type="submit"
+              class="btn bg-dark text-white">Edit Review
+            </button>
+          </div>
+        </form>
+      </div>
+    </div>
+  </div>
+</div>
+{% endblock content %}
