+++ moviesstore/movies/views.py
@@ -1,20 +1,17 @@
-from django.shortcuts import render, redirect, get_object_or_404
-@login_required
-def edit_review(request, id, review_id, Review=None):
-    review = get_object_or_404(Review, id=review_id)
-    if request.user != review.user:
-        return redirect('movies.show', id=id)
-    if request.method == 'GET':
-        template_data = {}
-        template_data['title'] = 'Edit Review'
-        template_data['review'] = review
-        return render(request, 'movies/edit_review.html',
-            {'template_data': template_data})
-    elif request.method == 'POST' and
-    request.POST['comment'] != '':
-        review = Review.objects.get(id=review_id)
-        review.comment = request.POST['comment']
-        review.save()
-        return redirect('movies.show', id=id)
-    else:
-        return redirect('movies.show', id=id)
+from django.shortcuts import render
+from .models import Movie
+
+def index(request):
+    template_data = {}
+    template_data['title'] = 'Movies'
+    template_data['movies'] = Movie.objects.all()
+    return render(request, 'movies/index.html',
+                  {'template_data': template_data})
+def show(request, id):
+    movie = Movie.objects.get(id=id)
+    template_data = {}
+    template_data['title'] = movie.name
+    template_data['movie'] = movie
+    return render(request, 'movies/show.html',
+                  {'template_data': template_data})
+
