--- moviesstore/moviesstore/moviesstore/movies/models.py
@@ -1,10 +1,0 @@
-from django.db import models
-class Movie(models.Model):
-    id = models.AutoField(primary_key=True)
-    name = models.CharField(max_length=255)
-    price = models.IntegerField()
-    description = models.TextField()
-    image = models.ImageField(upload_to='movie_images/')
-    def __str__(self):
-        return str(self.id) + ' - ' + self.name
-# Create your models here.
