+++ moviesstore/moviesstore/wsgi.py
@@ -4,7 +4,7 @@ WSGI config for moviesstore project.
 It exposes the WSGI callable as a module-level variable named ``application``.
 
 For more information on this file, see
-https://docs.djangoproject.com/en/5.1/howto/deployment/wsgi/
+https://docs.djangoproject.com/en/5.0/howto/deployment/wsgi/
 """
 
 import os
